requirements.txt
python-telegram-bot==20.7
moviepy==1.0.3